requirements.txt
fastapi
uvicorn[standard]
openrouter
pydantic
pydantic[email]
python-dotenv
PyJWT
python-multipart
passlib
bcrypt==4.0.1
scikit-learn
shap

